fix: iOS: Remove unused code
