--- ios/NativeSigner/Models/Seeds.swift
@@ -1,1 +1,0 @@
-
