+++ ios/NativeSigner/Modifiers/PrimaryTextField.swift
